Move binary addition/subtraction to ADD-TO-BINARY
The functionality was speculative to allow #{00} + 1

It made a lot of assumptions about overflow, endianness, etc. that seem
inappropriate for a fundamental operation...and should be part of a more
parameterized dialect.

For the moment, a theory is that NUL is represented as #{00} and not
as a strange state of ISSUE! (such as #"", which is more uniform if it
is treated as an empty string).  In the interest of experimentation,
the narrow math for NUL will be preserved instead of this.