Fix Back_Scan_Utf8_Char() corrupting size on error
Looking into a bug in FIND, it turns out that the size passed into the
scanning of a UTF-8 character was being changed in the case of returning
an error on seeing a 0 byte.

That didn't fix the bug--in fact, it turns out that's going to take more
time and a rewrite of some creaky old code.  But it did need to change.