Some refactoring work on TO and AS conversions
The first step of methodizing TO and AS got the code associated with
the type being converted from, vs. the type being converted to.

Originally the string handler did common work for things like String
and immutables like Issue, but this makes the more logical step of
taking the code with more concerns (Utf8 that may not be string
bearing) and makes that the common work.

Includes some other improvements, like validating EMAIL! and URL!
in TO and AS conversions (so you can no longer do things like
(to email! ""))