Re-unify native and intrinsic dispatchers
The first cut at dispatchers didn't pass a Level* in, because there was no
Level...so it stylized the interface to take the required OUT, ARG, and
a PHASE to pick out action-specific variables from the details array.

But a better way to do it is to pass a Level, but just with the
understanding that this is not always an Action_Executor()-level, and
it may be a parent which is being reused.  The OUT can still be used
as an output, and the argument can live in the SPARE.

This permits intrinsics to have refinements, and run efficiently when
those refinements are not in use.