Fix Sink corruption of callsite arguments
The way Sink() was written, it did corruptions of the target in the
constructor.  But when the target is a pointer to a variable, that may
be a variable on the stack being passed as another parameter.  So the
variable could be corrupted before it was passed.

The only way to avoid this is to defer the corruptions.