Add the ability to check datatypes for constraints
It was possible to ask things like ANY-UTF8? when you had an instance
of a type in your hand, like "abc" or [a b c].  But you couldn't check
the type itself, e.g. TEXT! or BLOCK!, to ask if it was in the
"type set" that the constraint represents.

In lieu of adding ANY-UTF8-TYPE? as a separate entry point, this uses
a refinement as ANY-UTF8?:TYPE.  That is a bit clearer that it's
distinct, whereas ANY-UTF8-TYPE? might sound a bit the same as
"ANY-UTF8?", like "is the type of the value any-utf8".  (This
confusion was present in historical Rebol, which called ANY-VALUE?
ANY-TYPE? to mean "A VALUE WITH ANY TYPE".