Allow matching of quoted quasiforms in FIND
The use of the Mold Cell Without Quotes function pointed out that the
quasiform status was being dropped.  This doesn't really make sense
in that the quasi is actually sort of part of the value in a way that
quoting is not (e.g. if you molded a ~ without its quote byte, you'd
get an entirely different form of _)

Also makes the QUOTE_BYTE() logic a bit clearer to emphasize that
the third quote level as ONEQUOTE_3 isn't the only single quoted
byte state...just the nonquasi one.