Swap UTF-8 based compare hook with string
The string hook was being passed non-node-bearing cells.  This instead
has the generic UTF-8 hook be delegated to by string.

Also has you pick ISSUE! out of ISSUE!, instead of giving back an
integer.  This annoyed earl historically, but it's less bad to pick
a single-codepoint issue out of an issue than a string from a string.