Add MAKE TUPLE! as transcoder
It used to be that TO TUPLE! of "1.2.3" would give you 1.2.3, but that
is not the philosophy of TO any longer.  Instead, TO TUPLE! of
"1 2 3" gives you 1.2.3

This adds MAKE TUPLE! "1.2.3" to pick up the missing functionality,
though this might be better as TRANSCODE:TYPE.

Also creates lists for numeric compressed tuples.