Auto-generate ELEMENT?, QUOTED?, QUASI? natives
This folds these tests in with the "Decider" based mechanics, so that
the optimization of being able to be packed into a parameter as a byte
can work.

ANY-ATOM? doesn't fit this model because it takes its argument meta,
and deciders don't do that.  But this needs to be revisited if things
like SPLICE? or PACK? are going to be auto-generated,