Rename MATCHER to TYPECHECKER, redefine MATCHER
MATCHER as a function generator sounds like something that would
behave essentially as a specializer for MATCH--e.g. return the
value passed on success.  The LOGIC!-bearing function seems
more like a typechecker.

The two codebases should be merged together and streamlined.
This takes the step of grouping the code together in the same
file, but more work needs to be done.