Allow intrinsics to RAISE or FAIL, show {fulfilling} frames
An initial rule for intrinsics was that they couldn't fail, because they
did not have stack levels to report in the backtrace.  But that's a
fairly impractical rule, compared to just making stack traces look for
the intrinsic flag and be aware of it.

So long as that code was being revisited, this tweaks the decision to
not show fulfilling frames in the WHERE backtrace to instead show them,
but put the level name in a FENCE!.  This is much better.