Move Level label into Action_Executor() Level state
Only Action levels have a label, so it shouldn't be in the Level struct,
but rather the union for Action_Executor() state.

This moves it, but doesn't move the debug string that gives the UTF8
of the label...because it's easier to debug view it from the Level
itself.

Also moves the expr_index into the evaluator state.  This had been used
by PARSE3 in SUBPARSE, which was an evaluator-like native...raising the
question of whether maybe it's more of a feed property.  But as it
wasn't being used anywhere that's just disabled for now.  Review.