Give all levels an extra SCRATCH cell
This makes the difficult decision to expand the size of all Level
structs by one cell.

It's based on the desire to be able to make any Level able to dispatch
an intrinsic, reusing its own Level.  Without another cell's space,
there's nowhere for an action executor to put the intrinsic (as the
SPARE is used for the argument).