Don't overwrite ->out cell in logic-based intrinsics
We want to be able to reuse levels in order to do intrinsic dispatch
for type checking, so that the Level of an action can be passed to
the typechecking native--and it can use the SPARE and the SCRATCH
for the argument and the action value of itself.

But that still leaves the question of multiplexing the output.  And
if the logic-bearing intrinsics write the output cell with a logic,
then they will be overwriting the result they are checking. :-(

This introduces a new BOUNCE_OKAY result, which can be checked for
along with nullptr.  Then the new rule is that logic-bearing
intrinsics should use these values and not overwrite the out cell.