Thread Levels through typechecking functions
Typechecking with functions started out very ad-hoc, with the functions
running stackfully, with their own stackful temporary cells.

The intent of intrinsics is that an existing Level be applied to calling
the typechecker, but that Level is having its SPARE and SCRATCH cells
used in a very specific way.  This needs a contract between the caller
and the typecheckers, so this makes it very obvious by giving the
routines names advertising the usage of spare and scratch.  Also the
functions will set the appropriate cells to unreadable in the debug
build on all code paths to reinforce that you can't count on them
being untouched.