Trampoline doesn't erase STATE_0 level outputs
An invariant established to support generators/yielders was that any
continuing Level would not have STATE_0 bytes.  While that code has
lapsed and isn't working at present, the invariant is a good one for
other reasons...along with having the output cell be erased on all
of these "initial entry" states.

However, enforcing it by having the trampoline do the erasing costs
a test of the Level's byte on every trampoline bounce, and also
may doubly-erase cells.  This switches it over to being an assert,
and pushes the responsibility over to Push_Level() and functions
that actively reuse a level.