Limit Probe() output, truncate binary molds
Due to the fact that Probe() on something like Lib would cause a very large
mold, it seemed like it was crashing to the C debugger because it took too
long to return.

This cuts down on the output to 10000 characters, and limits the molding
of binaries to help with that.