Change words like RETURN, CONTINUE, THROW to tripwires
Erroring functions are deceptive, because they still can be fetched
as functions...so it may be fairly far downstream before you realize
you didn't get what you wanted.

Making these tripwires instead means they will error if you try to
do a function fetch like RETURN/ and no function generator is
supplying it...instead of having to wait until if-and-when that
fetched function is ever used.