Simplify parameter list generation
Scanning a spec to produce a list of parameters has undergone a lot
of evolution, and it still needs work.

But one thing that made it complex early on was that there was an
idea that functions may have RETURN as a plain argument or local,
vs. having separate LAMBDA and FUNCTION types that differentiate it.

If you know up front that there's going to be RETURN, you can just
go ahead and push it in the first position and everything gets
a lot easier.  This goes ahead and does that, as the system depends
on it at the moment anyway.

(Questions about how to add an overriding RETURN on top of an
existing parameter list while still keeping the parameter order
for the functions being composed upon have not been sorted out,
but this should make that easier to reason about.)