Put primed cell inside the Evaluator Level
It became unavoidable that when multiple evaluator steps were taken
that the previous evaluation would have to be stored somewhere, in
order to facilitate invisibility.

One option would have been to add a cell to the frame, which would
have grown all Level types.  Non-Evaluators could use the cell for
other purposes.

But a different option of using the data stack was taken, pushing a
single element to hold the evaluations.  This was a bit of a trick,
but it seemed "cheaper".

Other changes made it so a cell's worth of space that had been
dedicated to the evaluator got moved out to a scratch cell that
is integral to intrinsics when being called from other levels.  This
means the evaluator wasn't using its space, so it might as well
use it for the primed cell.