Fix COUPLING OF accessor
It's no longer called a "BINDING" when a FRAME! cell has a pointer to
an object or frame in it (as definitional RETURN, YIELD, CONTINUE, etc.
do, and soon more functions will as well for the .member syntax).  It
is called a "COUPLING".

The accessor for extracting hadn't been renamed.