Rethink Fresh/Erased Cell strategy
In Ren-C, it's not legal to blindly overwrite the bits of a cell with a
new header... because some bits belong to the cell.  There are also debug
checks--such as that you're not overwriting a cell that is "protected".

So in the general case, any persistent bits (CELL_MASK_PERSIST) have to
be masked out before updating non-persistent bits in the header.

However, there are some spots in the system where persistent bits do
not apply...in particularly important places like the evaluator output
cell, spare cell, and scratch cell...as well as any evaluative target
(like the argument cells in a frame) which can be pointed to by the
output cell.

It would be nice if a very simple operation of writing a 0 to the
header bits could be used in these cases.  This brings safety to that
idea, and simplifies the model by removing "Fresh" cells from the
consciousness... only freshening when cells are being reset and
then moving them out of the fresh state.