Resurrect Generator and Yielder
In the initial implementation of stacklessness circa 2020, there was an
attempt to make a GENERATOR abstraction, built on a more generalized
YIELDER abstraction that could take arguments.

Because it wasn't central to the purposes, the tests weren't added to
the tests...and it atrophied.  This resurrects the code and gets it
in better shape for the modern system...though there are still many
questions raised by having detached stacks, with incomplete generators
having an arbitrary lifetime, and also potentially holding locks
on things they are enumerating.