Raised error warning suppression only on Move_Atom()
This rethinks how the assertion pertaining to non-intentional overwrites
of raised errors works.  Moving only concerns itself with blocking the
assertion when it's dealing with Atoms (as raised errors cannot exist
in Element or Value cells).  And it also makes it so that suppressing
errors doesn't canonize the state of a cell, to help ensure it can be
zero cost in the release build without that being a problem.