Move throw/longjmp into fail() macro
When using Clang static analyzer, it is the case that noreturn functions
are considered to abort(), and are thus not candidates for leak checks.

However, throw is potentially a candidate for leak checks.  Although it
does not work today, there's a path forward for it:

  https://github.com/llvm/llvm-project/issues/113399

In the hopes that will come to pass, this moves the `throw` or the
`LONGJMP` to the callsite of fail().