Replace METH with METHOD
METH originated from a time when FUNC and FUNCTION meant different
things (FUNCTION locally-gathered args, FUNC did not).  So member
functions needed the same distinction.

The difference was scrapped some time ago.  FUNC is still used as
an abbreviation, and also to help catch the stragglers that are
still using the deprecated FUNCTION.  But there's no particularly
good reason to keep METH around, as it's a bad abbreviation, and
there aren't METHOD stragglers in any appreciable number to be
worried about.