Limit inheritance of TOP_LEVEL feed const bit
On something like a `make object! <<block>>` situation, whether the
eval of that block should send a "wave of const" is decided by however
the block argument is fulfilled.  If you're in the body of a function
and you say:

   foo: func [] [make object! [foo: [] append foo 'x]]

Then what stops the APPEND is that the block was evaluated and captured
as const in the line of evaluation of a function body.  But if you say:

   foo: func [] [make object! copy [foo: [] append foo 'x]]

The MAKE OBJECT! is receiving a block that is free of that issue.  So
when the evaluation inside MAKE OBJECT! happens, it should be going
by the const of the block value it gets, not the const of the level.

However, when you're doing something like a GROUP! execution inside an
eval, you need to inherit:

   foo: func [] [(((let block: [a b c], append block 'd)))]

Each time a GROUP! is entered in needed to inherit the top level's
flag that says it's executing a feed that is const.