Allow .WORD to run actions
The choice to make .WORD the way you access member variables leads to
a question of if it's willing to buck the general behavior of TUPLE!
and let you run actions.

If it didn't, you would have to say /.word -or- /word would have to
imply picking out of the current object.

/word picking out of the current object would defeat other higher
priority notational usages, like /var: ... for assigning actions,
so it seems the better idea is to allow the exception that if you
have a word led by a dot then it will follow the same pattern as
without a dot and allow dispatch of actions undecorated.