Push further on "Coupling isn't binding"
The cell slot (Extra) which holds one platform pointer's worth of data
in each cell was at one point called "binding" for any time that it
contained a Node.  That was chipped away over time, but not completely
pushed through in the source.

This officially shifts it to where doing BINDING() on anything that
isn't a truly "bindable" type will assert.