TO PATH! of BLOCK! not used to make paths from parts
A behavior that was depended on pretty heavily was:

   >> to file! 'abc/def.ghi
   == %abc/def.ghi

When the various symmetries that are being applied for TO were taken
into account, this implies that TO PATH! of [abc def.ghi] can't be
`abc/def.ghi`, because that implies TO PATH! of "abc def.ghi" would
aslo be `abc/def.ghi`, which implies TO TEXT! of `abc/def.ghi` has
to be "abc def/ghi".

Untangling that means that JOIN has to take over the responsibilites
of TO...which is all right, because it makes more sense.

This starts work on taking over the code for DELIMIT for JOIN.