Rework JOIN to always use data stack, work on BLOB!
The original DELIMIT codebase would generate a Molder that was on the
stack during evaluations, and keep adding to it.  This presented some
problems at first with generators and yielders, because they needed
to yield across delimiting.

This was addressed by trying to put the Molder* into a HANDLE!.  But
then the question of how to deal with a BLOB! that doesn't use a
Molder came up.  And in general, it's just more complex to not make
use of the data stack and its automatic cleanup, in order to trap the
allocation of memory to free it.  It leaked, and needed review.

Here this takes the approach of going ahead and collecting all the
items for the JOIN on the data stack, even in the string case.  Then
it adds behavior for the BINARY!.

Note this changes when errors are delivered--which may be a good or
bad thing.  It could be seen as good if doing parallel implementations
that were based on reducing first, and then processing.