Add probe failures to JavaScript promises
The number of "uncooperative" abrupt failures (e.g. longjmp() or throw)
has been paring down, to where if it's possible to return a failing
result as a Bounce from a native it does.

However if the error was being bubbled up cooperatively in this way
to a JavaScript promise it was not being probed during ReplPad console
boot.  This adds a PROBE() for that case of needing to see the errors
before the console is working.