Make intrinsic arg typechecking fail fast
Intrinsics do their own typechecking, but if that typecheck fails then
it needs to be able to skip on to the next function if it's part of
a typecheck itself.  This makes that cheap and clear through a
special BOUNCE_XXX value specifically for intrinsics to say that
their arguments didn't typecheck.