Light reorganization of sys-cell.h
This tries to build things up in a little bit clearer manner.

It also makes it so that mutating bindings uses Assert_Cell_Writable(),
and Assert_Cell_Initable() goes through the check for raised errors...
which means Init_Unreadable() has it too.

Some superfluous code in the destructors for cell subclasses is
removed, because there's not a particularly good explanation for it
or what category of bugs that's supposed to have ever caught, as it's
not clear because they've never caught any.

Also renames CELL_MASK_0 to CELL_MASK_ERASED_0, just to drive home
more what an erased cell is.