Get rid of the "raised overwrite" check
This was a useful tool in the beginning of raised errors, but it's the
wrong shape for addressing the question of whether atoms containing
raised errors are handled correctly...which should come from some
management of the Atom type itself.  What winds up happening is that
you throw the suppression functions in at what seem like random
locations.

It's *definitely* not the long term solution to the problem, so this
gets rid of it to make space for a better answer...with the accepted
risk that maybe some raised errors would be swept under the rug
in the near term.  But those bugs aren't worse than any other bug,
vs. convoluting source in inexplicable ways for the wrong rule.