Eliminate PARAMCLASS_RETURN
The question of how functions give back RETURN specifications is
shifting to become dispatcher-specific.  The idea of there being
a particular PARAMCLASS_RETURN is no longer how it works.

This takes a big first step to rethinking it by eliminating the
idea of a particular PARAMCLASS for RETURN.  In the case of a
FUNC or NATIVE, they just happen to store their return in the
first cell of their paramlist...but other dispatchers may draw
the answer from other places.