Narrow search in Get_Context_From_Stack()
Climbing the stack looking for a native isn't a good way to decide where
to run code.  e.g. if the Stepper_Executor() is called from a native
living in a module, that Stepper's API calls should not see the module's
definitions.  So really, only the TOP_LEVEL should be taken into account
for expanding the search beyond lib.

Some comment cleanup in other %m-stacks.c functions.