Switch Get_Action_Flag() to Get_Phase_Flag()
It's kind of important in a low-level flag setting routine to know
what you're setting a flag on.  With there being a level of abstraction
over Action* as sometimes being a VarList* and sometimes being a
Phase*, if you think you're setting a flag on a single identity of
an Action* you might not understand that you're affecting other
identities if it's a VarList*.  This drops that abstraction.