Get rid of historical partials code
The idea of a partials array being separate from the frame, and then
"virtualizing" the varlist order--is abstract and bug prone.  It's a
tax on any code that tries to work with parameter lists.

Solutions to "refinement promotion" that don't try to virtualize the
process--but actually create new legitimate function interfaces--
are the direction being pursued now.