Eliminate KeySource conflation of Level/KeyList
It was decided that FRAME! needed to have an "adjunct" object.  This
became particularly necessary when FRAME! and ACTION! were unified
into one type.  This took up the MISC() slot in the Stub.

That had previously been able to be used for a place to put the
currently running Level, mapped to from a VarList.  Since that cheap
mapping was still necessary, the idea came up to swap out the keylist
of a running Varlist to be the Level, and then figure out the keys
from the running Phase.

This was confusing.  Instead, there's an acceptable limitation that
you can say running frames get their adjunct from their Phase (which
makes more sense).  Then running frames use their MISC() field for
the Level*.

It eliminates dependence on detection of Level* as being a NODE that's
not UNREADABLE and is a CELL.  That can potentially free up 3 flags
in the Level, assuming there's no other motivation to use this.