Make HAS_RETURN a details flag, not varlist flag
Beyond just the fact that varlist flags are scarce, the providing of a
RETURN type specification is now exclusively something that would be
based on the details dispatcher.  e.g. the Encloser_Dispatcher() is
what knows that the return type should come from the enclosing
function, and not the function being enclosed.  So it makes sense to
move this flag--if it should exist at all--onto the Details.