Don't react to L->out in Drop_Level()
This drops some code that was a bit dicey about automatically "rolling
back" the various bits of evaluator state that are tracked, by looking
at the throwing state or the value in the output cell.  Really that is
better managed explicitly, with places that purposefully want to
absolve code of the responsibility to balance things like the stack
or API calls (or the mold stack...)

Really this needs more thought and design, but it's an improvement
over what it was doing.