Kill off dummy typecheck in REFRAMER
It was already questionable, but RECYCLE:TORTURE discovered that it would
have to manage the frame cell passed (and it was using a frame that might
be incompatible with the function).  It's at the bottom of the priority
list to generate the error at creation time vs. call time, so drop it.