Update ENCLOSE to work with AUGMENT
ENCLOSE was not written to handle being passed an expanded frame, when
there was no associated phase.  But the concept for efficiency is that
you can do just that, e.g.

    /append-plus: enclose (
        augment append/ [extra [word!]]
    ) lambda [f] [
        f.extra  ; we expect this to be visible
    ]

This rigs up some rough behavior to handle that under the new model,
though all of this needs to be tightened up once things settle more.