Improve quasiform word init speed, and null/void/okay
There are preventions from creating antiforms of WORD! that aren't in
the accepted symbol set when you are running generic code.  However, the
OKAY/NULL/VOID symbols are known to be legal, so when internal code
wants to make those antiforms it shouldn't have to go through the
coerce code.

Similarly, there's prevention of creating quasiforms of arbitrary
types.  However, words are known to permit all quasiforms (only antiform
is limited).  So there's no need to go through that generic check
when initializing quasi-words.