"Blit" parameter cells
There was some question over whether doing a memcpy() of a ParamList into
the cells of a frame being built would be advantageous, vs. trying to
erase cells one at a time and fill them.  The thought was that despite the
extra "walk" the memcpy() implied, it might be faster.  It wasn't.

What did improve things was switching to copying the cells without masking.
This actually makes the code a little clearer than CELL_MASK_COPY_PARAM
was making it.