DECLARE_ELEMENT() etc. use erased cells
When CELL_MASK_UNREADABLE came along to allow for unreadable cells,
there was a thought this needed to be used with Push_Lifeguard() so
it could always know it was guarding something with NODE_FLAG_NODE.

This was kind of silly, since evaluation can erase those cells and
has to tolerate the zero state.  So blocking them on Push_Lifeguard()
only means it costs a little bit more (potentially) to assign the
cell header, vs. an immediate 0 value.