Bootable move of Exemplar to "archetype" phase (WIP)
There were conflicting needs of the "archetype" slot (maybe better
called "rootvar").  One was to have an ever-handy representation of
a Cell for an object that existed in-memory and could be used any
time you had only a Stub pointer to a VarList but wanted a Cell*.
This required the VarList slot inside that cell to have the same
pointer as the VarList itself (hence redundant, and "wasting space").
The other was to provide additional information which was not
available a Cell referring to the VarList--essentially offering
2 more pointers as a slighly-less-localized augmentation of the
slots in the Stub itself.

These needs were contradictory, and in practice the strange cells
in the initial slot introduced too many deviations to be thought of
as proxies for the cell itself.  This starts the move away from
that use of the "archtype" slot, and instead goes fully into the
"additional information" concept.  So FRAME! archetypes now hold
decoupled FRAME! cells which can have any "phase" desired.

This partially-working commit is being made while naming shifts to
change the alternate "Phase_Or_Label" slot to "Lens_Or_Label" and
have the Lens always be a ParamList*, in order to help clarify,
as this has become very jumbled.