Shrink ENCLOSE and ADAPT Details arrays
With the new concept that Details[0] does not need to hold a self
referencing Details*, that slot now can act as a holder for a
provider of interface--whether that be a direct provider, or a
Details-based frame that indirectly can lead to a ParamList.

Consequently, functions like ENCLOSE and ADAPT don't need to
store a separate copy of the action that provides the interface,
they can just use what's in the Details[0] slot.