Go deeper for level labels, init with lensing
The use of lenses on plain frames in order to maintain the display
of their fields even after specializing values is contentious with
those frames being able to hold labels for the action they were
created for.  This rigs up a rickety choice to extract the label
deeply (e.g. from the archetype).  It also makes the initialization
of lensed frames a special form of init that doesn't take a label,
so it doesn't misleadingly initialize with a label then tweak the
lens afterward...making lensed frame initializations easier to find.