Eliminate antiform parameters ("holes")
Naming and dealing with antiform parameters caused several issues,
which had new solutions available after they became legal in
FRAME! to be passed as arguments--but only illegal to be used in
frame specialization slots.  Once that decision was made, the fact
that the pervasive use of PARAMETER! was its antiform meant that
using antiforms at all just raised a question of what to call that
antiform, when "parameter" was the natural name.  This gets rid of
the idea of using antiforms at all, accepting the consequence that
not everything you can put in a BLOCK! can be used in specialization
slots...PARAMETER! is an exception.  But it's not entirely clear
what that rule was buying, considering all the confusion that the
antiform caused.