Move RETURN to Details of Natives
When RETURN was first introduced, the type information was kept in the
frame definition because there was a slot needed for it on invocation
to hold the definitional return action.  So it was an effective use of
space for a slot that would be otherwise unused.

The same isn't true for natives, which don't need a definitional return
to "return from themselves".  So the slot was wasting space in each
instantiation.  Some natives took advantage of the GC available slot,
but few needed it, when SCRATCH and SPARE and OUT are available.

This shifts it so that natives no longer put RETURN in their frames.
Instead they keep their return types in their Details, so one slot per
native.  The uses of ARG(return) are adjusted to use local variables
which can now be allocated just in those few natives that need more
GC-protected slots.

Several improvements are made in the process here, including that
YIELD now can have a return type defined for it in YIELDER (like
RETURN), and the PARAMLIST_HAS_RETURN flag is no longer relevant.