Protect against multiple uses of NextVirtual slot
There was an idea that the LINK() slot in a VarList representing a frame
or an object could be used to point to items in a virtual link chain,
so you wouldn't need to make a separate USE Stub.

But this raises a question of interpretation--what Lens should that
imply, if the archetype slot of a running frame changes to reflect the
current frame?  Really it only makes sense if it's the last one.

Also fixes the JavaScript module in the process.