Forget double-dereference on LIBREBOL_BINDING
There was an unimplemented concept that it might be possible for a
LET statement or other operation to actually change the context after
an API call, for example:

    rebElide("let x: 100");  // imagine it could update "binding_ref"
    assert(100 = rebUnboxInteger("x"));  // could see new value

This would open a can of worms related to usermode code holding non-GC
visible references to LET statements.  The only way to deal with it
would be if there were APIs for manipulating the contexts, and they are
invisible in the API by design.  If they become visible, they're just
context Value* that you pipe around manually.

So this kills the double-dereference idea, simplifying things and
giving clarity by eliminating something that wasn't going to work.