Eliminate Get_Context_From_Stack()
This gets rid of a bad function that added runtime cost and muddied
the waters.  It was based on trying to detect what native was running
and get the context from that.

The better solution is to fully embrace the idea that the API needs a
LIBREBOL_BINDING_NAME to associate with.  Modules now legitimately set
a global variable that gets shadowed in their natives, and if code is
run in a service routine it will see that global variable unshadowed.
No stack climbing required!