Make TCC use Api_Function_Dispatcher() (untested)
Prior to the shadowed binding idea, there was a concept that things
that considered themselves to be "natives" would obey a certain layout
in their Details array, so that stack climbing could extract a context
from them to know where to bind code.

That idea is gone, and now the concept is that if you want to use the
API inside a native then you use the Api_Function_Dispatcher() and it
does the right thing.  But if you don't use that dispatcher, you have
to have access to the low level API to do Varlist tinkering if you
expect things to work.

So now the IS_NATIVE flag really only applies to *raw* natives, which
have to answer things like what their return is and what their body
is without having a registration of the dispatcher.  This makes that
completely clear that it has this narrow function, and migrates TCC
to producing ordinary API natives.

It is currently untested, because there are bigger things to worry
about--but this moves in the right direction by getting the confusing
flag off of places it doesn't belong.