Add beginnings of all-ASCII string optimization
There was a mistaken belief at one time that whether a UTF-8 string was all
ASCII would have to be tracked by a flag that could become outdated.  But
since you can simply test if the size in bytes matches the number of
codepoints, this is not the case.

This begins that, including fixes to the DEBUG_UTF8_EVERYWHERE build.