Eliminate node_MISC() Hitch complexities
The MISC() and LINK() etc macros had a somewhat rigid structure that
didn't fit the case when more than one flavor of Stub could hold the
same property, which led to awkwardness.

This gets rid of that awkwardness by using more normal inline functions
to do the same job.  Which is probably better than the convoluted
mechanism that was invented for the sake of preserving assignments.