Get rid of BINDING() macro and BindingHolder template
There was an idea to make it so you could use a macro that expanded
to a C++ class to do additional checking for:

    BINDING(c) = something;
    Context* binding = BINDING(c);

This was thought to convey more that it was "simple" than having to
do the get and fetch with separate function-looking things.

   Tweak_Cell_Binding(c, something);
   Context* binding = Cell_Binding(c);

But the latter has advantages, including being able to do its checking
in non-C++ build.  But also not having to explain the C++ code that
does the weird behaviors.

This switches to the simpler approach.