Get rid of PAYLOAD() and EXTRA() macros
There was a semi-interesting idea in the early development which
was that you would be able to write C code that would pick fields
out of structures and then automatic checking from C++ would be
able to add extra behavior to the code.

Hence instead of saying:

   assert(Cell_Heart(cell) == REB_INTEGER);
   cell->payload.i64 = 100;

You would say:

   PAYLOAD(Integer, cell).i64 = 100

And the checking would be done automatically for you.

This never materialized, and actually in the places where it did
materialize (e.g. BINDING(cell)) it wound up being walked back,
just because it's clearer to have inline functions like:

    Tweak_Cell_Integer(cell, 100);

Than to rely on C++ magic, that is hard to design to be general
and good... and if it's generalized it is hard to put strategic
breakpoints exactly where you want them anyway.

Since no value emerged, this rolls back the idea completely.