CELL_NODE1, CELL_NODE1, CELL_EXTRA
In reevaluating the use of LValue semantics for certain assignments,
the actual "baby in the bathwater" is the ability to declare in one
place how the slots of a cell or stub are used.  This makes some
definitions for cells so you can do that...though it is a bit
messy for starters.  But the ultimate goal would be to have each
cell and stub declaring its usage of slots up front.