STUB_LINK(), STUB_MISC(), etc.
This does a more straightforward reclaiming of the idea behind the LINK()
and MISC() macros, which was to make a clear centralization of how each
Stub used its slots.

Under this new idea, the macros aren't trying to do any magic--they are
just aliases for the node inside the stub.  So you still have to cast
them to use them.  But it makes it easier to get an overview of the
slot usage, which can be confusing...and makes it easier to shuffle
them around when necessary.