Restyle USE Stubs so caller fills
This switches it so that a Use stub gets created so there's space to
initialize the Cell of what to use, vs. needing to construct a Cell to
pass in.

A downside is that this loses the ability to check the Cell in the stub
is a valid thing for the Use, but that can be done by GC.  TBD.