%sys-tick.h header file, Probe_Limit()
Explaining the concept of ticks and how they are used in the system,
plus the definitions related to them, deserves its own header file.
(%sys-trampoline.h was being used, but it doesn't make much sense
to include that higher-level file before %sys-node.h, and so the
tick definitions in %sys-tick.h makes even more sense).

Also, PROBE() was being chopped off by default when it's not a
good default.  This adds Probe_Limit().