Eliminate generic REDO function
The idea of REDO being something that could work with a Dispatcher
for an in-flight function in a generic way--without participation of
that dispatcher--was a broken idea.

Were such a thing to be possible, the idea would have to be that
Dispatchers advertised their ability to intercept these returns and
do the right thing.  There's not currently a clamoring demand for
that protocol, so this narrows the existing tail call scenarios to
just RETURN:RUN as a proof of concept.