Hack in SUBPARSE:MATCH, SUBPARSE:RELAX, kill VALIDATE
VALIDATE was a word chosen prior to the existence of PARSE:MATCH for
a combinator that would run parse rules and then give back the input
instead of the parse result.

Then, once :MATCH was firmed up, it wasn't possible to do refinements
in UPARSE on combinators yet.

A bug in VALIDATE motivated going ahead and hacking in the variations
via literal combinator match, as it should be possible to use
refinements with combinators in the future.  (Whether someone wants
to shorthand this as VALIDATE or not would be up to them.)