Eliminate ~end~ antiform, refinements NULL in MAKE FRAME!
Distinguishing required fields you haven't assigned in MAKE FRAME!
from "reaching the end" was analyzed as not being worth it.

Also, under the new line of thinking that only ^META parameters can
receive nothing antiforms (and they receive them as meta-nothing,
a.k.a. quasiform blank) this made plain nothing the obvious choice
for the "no parameter" state that the `<end>` flag allows.

(`<end>` itself should probably be renamed, but this is still being
debated as to what to call it)

In order for all the ducks to be in a row with this, the behavior
that ~ antiforms would automatically turn into null for refinement
slots was moved from the typechecking phase (where it never seemed
correct) into the "make frame" phase.  This meant makes it possible
to distinguish between a refinement that has been promoted (e.g.
`make frame append:dup/` would promote DUP, hence it should default
to a nothing) vs. a refinement that had not been promoted (in that
example, e.g. :PART and :LINE, which would be nulled)